demo/lancedemo/ecommerce/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
lancedb==0.5.0
sentence-transformers==2.2.2
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.14.0
numpy>=1.21.0
pandas>=1.3.0
tokenizers>=0.13.0
Pillow>=8.0.0
requests>=2.25.0